vpn-admin/requirements.txt
Django==3.2.7
django-environ==0.7.0
django-filter==21.1
django-autoslug==1.9.8
django-countries==7.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
django-phonenumber-field==5.2.0
django-cors-headers==3.13.0
phonenumbers==8.12.33
django-import-export==2.8.0
django-money==3.0.0
sentry-sdk==1.9.5
requests==2.28.1
qrcode==7.3.1
psycopg2-binary==2.9.1
pymorphy2==0.9.1
gunicorn==20.1.0
docker
Pillow==9.2.0
uritemplate==4.1.1
python-dateutil==2.8.2